office-mate-backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pytesseract
Pillow
PyPDF2
scikit-learn
joblib
